App container ready. Rendering List view placeholder
